% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{apply_height_eq}
\alias{apply_height_eq}
\title{Apply height equation}
\usage{
apply_height_eq(
  dbh,
  site_index,
  top_dob,
  stand_basal_area,
  b1,
  b2,
  b3,
  b4,
  b5,
  b6
)
}
\arguments{
\item{dbh}{vector of DBH observations in inches}

\item{site_index}{vector of site index estimates (base age 50)}

\item{top_dob}{vector of outside-bark diameter values to which height will
be estimated}

\item{stand_basal_area}{vector of stand-level basal area estimates
corresponding to the stems}

\item{b1, b2, b3, b4, b5, b6}{numeric vector of model coefficients from Table 1}
}
\value{
vector of height estimates
}
\description{
Apply the height equation (eq. 2 in publication) from RP NC-250 using
provided dbh, site index, outside-bark top diameter, stand basal area, and
coefficients from Table 1 in the publication.
}
\examples{
height_eq(
  dbh = 10,
  site_index = 60,
  top_dob = 4,
  stand_basal_area = 75,
  b1 = table1$b1[1],
  b2 = table1$b2[1],
  b3 = table1$b3[1],
  b4 = table1$b4[1],
  b5 = table1$b5[1],
  b6 = table1$b6[1]
)
}
