% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{apply_small_tree_biomass_eq}
\alias{apply_small_tree_biomass_eq}
\title{Apply small tree biomass equation}
\usage{
apply_small_tree_biomass_eq(dbh)
}
\arguments{
\item{dbh}{vector of DBH observations in inches}
}
\value{
vector of biomass estimates in tons
}
\description{
Apply the equation used for estimating biomass in trees less than 5" DBH
}
\examples{
\dontrun{
apply_small_tree_biomass_eq(4)
}
}
