% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{apply_volume_eq}
\alias{apply_volume_eq}
\title{Apply volume equation}
\usage{
apply_volume_eq(dbh, height, b0, b1)
}
\arguments{
\item{dbh}{vector of DBH observations in inches}

\item{height}{vector of heights to which to estimate volume}

\item{b0, b1}{vector of model coefficients from Table 2 or Table 3}
}
\value{
vector of volume estimates
}
\description{
Apply the volume equation (eq. 1 in publication) from RP NC-250 using
provided dbh, height to merchantable limit, and coefficients from Table 1 in
the publication.
}
\details{
The volume equation is as follows:
\deqn{V = b_0 + b_1 D^2 H}

Where:
\itemize{
\item \eqn{V} is the volume in either cubic feet or board feet
\item \eqn{b_0} and \eqn{b_1} are species specific coefficients from Table
1
\item \eqn{D} is diameter at breast height in inches
\item \eqn{H} height to which volume is to be computed
}
}
\examples{
\dontrun{
apply_volume_eq(
  dbh = 10,
  height = 40
  b0 = table2$b0[1],
  b1 = table2$b1[1]
)
}
}
