% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_species_group}
\alias{assign_species_group}
\title{Assign species group}
\usage{
assign_species_group(spcd)
}
\arguments{
\item{spcd}{vector of FIA species codes}
}
\value{
character vector of valid species groups from the tables in the
original publication
}
\description{
Assign individual FIA species codes to one of the 'Species Groups' as
described in the original paper.
}
\examples{
\dontrun{
  assign_species_group(c(541, 371, 95, 73))
}
}
