% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{estimate_bark_weight}
\alias{estimate_bark_weight}
\title{Estimate bark weight}
\usage{
estimate_bark_weight(gross_vol_cuft, stump_vol_cuft, bark_correction_factor)
}
\arguments{
\item{gross_vol_cuft}{estimates of gross volume in cubic feet}

\item{stump_vol_cuft}{estimates of stump volume in cubic feet}

\item{bark_correction_factor}{species-specific bark correction factors
estimated by \code{\link{get_bark_correction_factor}}}
}
\value{
estimates of bark weight in pounds
}
\description{
Estimate bark weight in lbs following equation 5 in the original publication.
}
