% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{estimate_biomass}
\alias{estimate_biomass}
\title{Estimate biomass}
\usage{
estimate_biomass(
  spcd,
  dbh,
  height = NULL,
  site_index = NULL,
  top_dob = NULL,
  stand_basal_area = NULL
)
}
\arguments{
\item{spcd}{vector of FIA species codes}

\item{dbh}{vector of DBH observations in inches}

\item{height}{optional vector of heights to use for volume computations. If
not provided, heights will be estimated using supplied values for
\code{site_index}, \code{top_dob}, and \code{stand_basal_area} using
\code{\link{estimate_height}}}

\item{site_index}{vector of site index estimates (base age 50)}

\item{top_dob}{vector of outside-bark diameter values to which height will
be estimated}

\item{stand_basal_area}{vector of stand-level basal area estimates
corresponding to the stems}
}
\value{
estimates of biomass in tons
}
\description{
Estimate green tons of biomass to a given top diameter following the system
outlined in the original publication. Biomass of the bole (stem + stump),
bark, and top are summed to obtain an estimate of the total biomass.
}
\examples{
estimate_biomass(
  spcd = c(541, 371, 95, 73),
  dbh = c(10, 11, 12, 13),
  site_index = c(60, 65, 60, 70),
  top_dob = 4,
  stand_basal_area = 80
)

# get biomass with pre-specified heights (i.e. total height observations)
estimate_biomass(
  spcd = c(541, 371, 95, 73),
  dbh = c(10, 11, 12, 13),
  height = c(50, 55, 60, 65)
)
}
