% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{estimate_bole_weight}
\alias{estimate_bole_weight}
\title{Estimate bole weight}
\usage{
estimate_bole_weight(spcd, gross_vol_cuft, stump_vol_cuft, bark_weight_lbs)
}
\arguments{
\item{spcd}{vector of FIA species codes}

\item{gross_vol_cuft}{estimates of gross volume in cubic feet}

\item{stump_vol_cuft}{estimates of stump volume in cubic feet}

\item{bark_weight_lbs}{estimates of bark weight in pounds estimated by
\code{\link{estimate_bark_weight}}}
}
\value{
estimates of bole weight in tons
}
\description{
Estimate bole (stump + stem) weight in tons following equation 6 in the
original publication.
}
