% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{estimate_height}
\alias{estimate_height}
\title{Estimate height}
\usage{
estimate_height(spcd, dbh, site_index, top_dob, stand_basal_area)
}
\arguments{
\item{spcd}{vector of FIA species codes}

\item{dbh}{vector of DBH observations in inches}

\item{site_index}{vector of site index estimates (base age 50)}

\item{top_dob}{vector of outside-bark diameter values to which height will
be estimated}

\item{stand_basal_area}{vector of stand-level basal area estimates
corresponding to the stems}
}
\value{
vector of height estimates
}
\description{
Estimate height to a given top diameter given species, diameter, site index,
and stand basal area. The height equation is applied by the function
\code{\link{apply_height_eq}}.
}
\examples{
estimate_height(
  spcd = c(541, 371, 95, 73),
  dbh = c(10, 11, 12, 13),
  site_index = c(60, 65, 60, 70),
  top_dob = c(4, 4, 9, 10),
  stand_basal_area = 80
)
}
