% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{estimate_stump_volume}
\alias{estimate_stump_volume}
\title{Estimate stump volume}
\usage{
estimate_stump_volume(spcd, dbh)
}
\arguments{
\item{spcd}{vector of FIA species codes}

\item{dbh}{vector of DBH observations in inches}
}
\value{
estimates of stump volume in cubic feet
}
\description{
Estimate stump volume following equation 3 in the original publication.
}
