% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{estimate_top_weight}
\alias{estimate_top_weight}
\title{Estimate top weight}
\usage{
estimate_top_weight(spcd, bark_weight_lbs, gross_vol_cuft)
}
\arguments{
\item{spcd}{vector of FIA species codes}

\item{bark_weight_lbs}{estimates of bark weight in pounds estimated by
\code{\link{estimate_bark_weight}}}

\item{gross_vol_cuft}{estimates of gross volume in cubic feet}
}
\value{
estimates of bole weight in tons
}
\description{
Estimate top weight in tons following equation 7 in the original publication.
}
