% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{estimate_volume}
\alias{estimate_volume}
\title{Estimate volume}
\usage{
estimate_volume(spcd, dbh, height, vol_type = c("cuft", "bdft"))
}
\arguments{
\item{spcd}{vector of FIA species codes}

\item{dbh}{vector of DBH observations in inches}

\item{height}{vector of heights to which to estimate volume}

\item{vol_type}{units for volume estimates, either \code{cuft} or
\code {bdft}}
}
\value{
vector of volume estimates
}
\description{
Estimate volume to specified height along stem in either cubic feet or
International 1/4 board feet. The equation is applied by the function
\code{\link{apply_volume_equation}}
}
\examples{
estimate_volume(
  spcd = c(541, 371, 95, 73),
  dbh = c(10, 11, 12, 13),
  height = 40, 40, 24, 24,
  vol_type = "cuft"
)
}
