% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{get_bark_correction_factor}
\alias{get_bark_correction_factor}
\title{Get bark correction factor}
\usage{
get_bark_correction_factor(spcd, dbh)
}
\arguments{
\item{spcd}{vector of FIA species codes}

\item{dbh}{vector of DBH observations in inches}
}
\value{
species-specific bark correction factor values
}
\description{
Compute bark weight correction factor based on species and DBH following
equation 4 in the original publication.
}
