% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_species}
\alias{ref_species}
\title{ref_species}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2677 rows and 79 columns.
}
\source{
data-raw/ref_species.R
}
\usage{
ref_species
}
\description{
ref_species table from USFS FIADB: see page 9-19 of FIADB 2.0 User Guide for
description
}
\keyword{datasets}
