% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{species}
\format{
dataframe:
\describe{
\item{species_group}{species group label from paper}
\item{scientific_name}{Genus species}
\item{common_name}{common name}
\item{spcd}{numeric FIA species code}
}
}
\source{
data-raw/species.R
}
\usage{
species
}
\description{
A dataframe relating the 'Species Group' in RP NC-250 to records in FIA's
ref_species table
}
\keyword{datasets}
