% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table1}
\alias{table1}
\title{table1}
\format{
dataframe:
\describe{
\item{species_group}{species group label from paper}
\item{b1}{first height model coefficient for diameter term}
\item{b2}{second height model coefficient for DBH term}
\item{b3}{third height model coefficient for diameter term}
\item{b4}{height model coefficient for site index term}
\item{b5}{height model coefficient for top diameter term}
\item{b6}{height model coefficient for basal area term}
\item{std_err}{standard error (feet) of model estimates for each species}
}
}
\source{
data-raw/tables.R
}
\usage{
table1
}
\description{
A dataframe containing coefficients for the height model equation from Table
1 in the original publication.
}
\keyword{datasets}
