% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{apply_height_eq}
\alias{apply_height_eq}
\title{Apply height equation}
\usage{
apply_height_eq(
  dbh,
  site_index,
  top_dob,
  stand_basal_area,
  b1,
  b2,
  b3,
  b4,
  b5,
  b6
)
}
\arguments{
\item{dbh}{vector of DBH observations in inches}

\item{site_index}{vector of site index estimates (base age 50)}

\item{top_dob}{vector of outside-bark diameter values to which height will
be estimated}

\item{stand_basal_area}{vector of stand-level basal area estimates
corresponding to the stems}

\item{b1, b2, b3, b4, b5, b6}{numeric vector of model coefficients from Table 1}
}
\value{
vector of height estimates
}
\description{
Apply the height equation (eq. 2 in publication) from RP NC-250 using
provided dbh, site index, outside-bark top diameter, stand basal area, and
coefficients from Table 1 in the publication.
}
\details{
The height equation is as follows:
\deqn{Height = 4.5 + b_1(1-e^{(-b_2D)})^{b_3}*S^{b_4}*T^{b_5}*B^{b_6}}

Where:
\itemize{
\item \eqn{b_1}, \eqn{b_2}, \eqn{b_3}, \eqn{b_4}, \eqn{b_5}, \eqn{b_6} are
species specific coefficients from Table 1
\item \eqn{D} is diameter at breast height in inches
\item \eqn{S} is site index (base age 50)
\item \eqn{T} is \eqn{1.00001 - d/D}, \eqn{d} is top diameter outside bark
\item \eqn{B} is stand basal area in square feet per acre
}
}
\examples{
\dontrun{
apply_height_eq(
  dbh = 10,
  site_index = 60,
  top_dob = 4,
  stand_basal_area = 75,
  b1 = table1$b1[1],
  b2 = table1$b2[1],
  b3 = table1$b3[1],
  b4 = table1$b4[1],
  b5 = table1$b5[1],
  b6 = table1$b6[1]
)
}
}
